+++ t/manual_reconnect.t
@@ -27,6 +27,7 @@ sub fresh_device {
         TEST_OPEN => 1,
         helper => {
             authenticated => 1,
+            lifecycleState => 'connected',
             protocol => 2,
             deviceType => 'wattpilot_flex',
         },